requirements.txt
-i https://pypi.org/simple
-e git+https://github.com/rji-futures-lab/django-postgres-copy@adb679d986ac3f728d7ef68073d1099e28c6e62d#egg=postgres-copy
argcomplete==1.9.3
boto3==1.9.174
botocore==1.12.174
certifi==2019.6.16
cfn-flip==1.2.1
chardet==3.0.4
click==7.0
django-bootstrap4==0.0.8
django-debug-toolbar==2.0
django-phonenumber-filter==1.0.0
django-storages==1.7.1
django-tastypie==0.14.2
django==2.2.2
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
phonenumbers==8.10.13
pillow==6.0.0
placebo==0.9.0
psycopg2-binary==2.8.3
python-dateutil==2.6.1 ; python_version >= '2.7'
python-mimeparse==1.6.0
python-slugify==1.2.4
pytz==2019.1
pyyaml==5.1.1
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.7
unidecode==1.1.1
urllib3==1.25.3
werkzeug==0.15.4
wheel==0.33.4
wsgi-request-logger==0.4.6
zappa==0.48.2
